Extract `ApiKeyDetailContent` so it can be reused
