+++ web/src/pages/project/[projectId]/traces/setup.tsx
@@ -6,7 +6,8 @@ import ContainerPage from "@/src/components/layouts/container-page";
 import { ActionButton } from "@/src/components/ActionButton";
 import { SubHeader } from "@/src/components/layouts/header";
 import { Button } from "@/src/components/ui/button";
-import { ApiKeyRender } from "@/src/features/public-api/components/CreateApiKeyButton";
+import { ApiKeyDetailContent } from "@/src/features/public-api/components/ApiKeyDetailContent";
+import { useLangfuseBaseUrl } from "@/src/features/public-api/hooks/useLangfuseEnvCode";
 import { type RouterOutput } from "@/src/utils/types";
 import { useState } from "react";
 import { useQueryProject } from "@/src/features/projects/hooks";
@@ -18,6 +19,7 @@ export const TracingSetup = ({
   projectId: string;
   hasTracingConfigured?: boolean;
 }) => {
+  const baseUrl = useLangfuseBaseUrl();
   const [apiKeys, setApiKeys] = useState<
     RouterOutput["projectApiKeys"]["create"] | null
   >(null);
@@ -42,10 +44,13 @@ export const TracingSetup = ({
       <div>
         <SubHeader title="1. Get API keys" />
         {apiKeys ? (
-          <ApiKeyRender
-            generatedKeys={apiKeys}
-            scope={"project"}
+          <ApiKeyDetailContent
+            scope="project"
+            secretKey={apiKeys.secretKey}
+            publicKey={apiKeys.publicKey}
+            baseUrl={baseUrl}
             className="mt-4"
+            showMcpSection={false}
           />
         ) : (
           <div className="flex flex-col gap-4">
